requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.9.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
requests==2.32.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.8.3
huggingface-hub==0.25.2
cohere==5.11.0
jinja2==3.1.4
